moved asset tracker def to utils.py
