+++ tests/system/python/e2e/test_e2e_coap_PI.py
@@ -44,18 +44,6 @@ def get_statistics_map(foglamp_url):
     jdoc = json.loads(r)
     return utils.serialize_stats_map(jdoc)
 
-def get_asset_tracking_details(foglamp_url, event=None):
-    _connection = http.client.HTTPConnection(foglamp_url)
-    uri = '/foglamp/track'
-    if event:
-        uri += '?event={}'.format(event)
-    _connection.request("GET", uri)
-    r = _connection.getresponse()
-    assert 200 == r.status
-    r = r.read().decode()
-    jdoc = json.loads(r)
-    return jdoc
-
 
 def _verify_egress(read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries, asset_name):
     retry_count = 0
@@ -151,7 +139,7 @@ class TestE2E_CoAP_PI:
         if not skip_verify_north_interface:
             _verify_egress(read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries, asset_name)
 
-        tracking_details = get_asset_tracking_details(foglamp_url, "Ingest")
+        tracking_details = utils.get_asset_tracking_details(foglamp_url, "Ingest")
         assert len(tracking_details["track"]), "Failed to track Ingest event"
         tracked_item = tracking_details["track"][0]
         assert "coap" == tracked_item["service"]
@@ -159,7 +147,7 @@ class TestE2E_CoAP_PI:
         assert "coap" == tracked_item["plugin"]
 
         if not skip_verify_north_interface:
-            egress_tracking_details = get_asset_tracking_details(foglamp_url,"Egress")
+            egress_tracking_details = utils.get_asset_tracking_details(foglamp_url,"Egress")
             assert len(egress_tracking_details["track"]), "Failed to track Egress event"
             tracked_item = egress_tracking_details["track"][0]
             assert "NorthReadingsToPI" == tracked_item["service"]
