+++ tests/system/python/e2e/test_e2e_coap_pi_web_api.py
@@ -48,19 +48,6 @@ def get_statistics_map(foglamp_url):
     return utils.serialize_stats_map(jdoc)
 
 
-def get_asset_tracking_details(foglamp_url, event=None):
-    _connection = http.client.HTTPConnection(foglamp_url)
-    uri = '/foglamp/track'
-    if event:
-        uri += '?event={}'.format(event)
-    _connection.request("GET", uri)
-    r = _connection.getresponse()
-    assert 200 == r.status
-    r = r.read().decode()
-    jdoc = json.loads(r)
-    return jdoc
-
-
 def _verify_egress(read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries, asset_name):
     retry_count = 0
     data_from_pi = None
@@ -174,7 +161,7 @@ class TestE2E_CoAP_PI_WebAPI:
         assert {DATAPOINT: DATAPOINT_VALUE} == retval[0]["reading"]
 
         if not skip_verify_north_interface:
-            egress_tracking_details = get_asset_tracking_details(foglamp_url, "Egress")
+            egress_tracking_details = utils.get_asset_tracking_details(foglamp_url, "Egress")
             assert len(egress_tracking_details["track"]), "Failed to track Egress event"
             tracked_item = egress_tracking_details["track"][0]
             assert "NorthReadingsToPI_WebAPI" == tracked_item["service"]
