+++ tests/system/python/e2e/test_e2e_csv_PI.py
@@ -51,18 +51,6 @@ def get_statistics_map(foglamp_url):
     jdoc = json.loads(r)
     return utils.serialize_stats_map(jdoc)
 
-def get_asset_tracking_details(foglamp_url, event=None):
-    _connection = http.client.HTTPConnection(foglamp_url)
-    uri = '/foglamp/track'
-    if event:
-        uri += '?event={}'.format(event)
-    _connection.request("GET", uri)
-    r = _connection.getresponse()
-    assert 200 == r.status
-    r = r.read().decode()
-    jdoc = json.loads(r)
-    return jdoc
-
 
 @pytest.fixture
 def start_south_north(reset_and_start_foglamp, add_south, start_north_pi_server_c, remove_data_file,
@@ -176,7 +164,7 @@ class TestE2E_CSV_PI:
         if not skip_verify_north_interface:
             _verify_egress(read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries, asset_name)
 
-        tracking_details = get_asset_tracking_details(foglamp_url, "Ingest")
+        tracking_details = utils.get_asset_tracking_details(foglamp_url, "Ingest")
         assert len(tracking_details["track"]), "Failed to track Ingest event"
         tracked_item = tracking_details["track"][0]
         assert "play" == tracked_item["service"]
@@ -184,7 +172,7 @@ class TestE2E_CSV_PI:
         assert "playback" == tracked_item["plugin"]
 
         if not skip_verify_north_interface:
-            egress_tracking_details = get_asset_tracking_details(foglamp_url,"Egress")
+            egress_tracking_details = utils.get_asset_tracking_details(foglamp_url,"Egress")
             assert len(egress_tracking_details["track"]), "Failed to track Egress event"
             tracked_item = egress_tracking_details["track"][0]
             assert "NorthReadingsToPI" == tracked_item["service"]
