+++ tests/system/python/e2e/test_e2e_filter_fft_threshold.py
@@ -53,17 +53,6 @@ class TestE2eFilterFFTThreshold:
         jdoc = json.loads(r)
         return utils.serialize_stats_map(jdoc)
 
-    def get_asset_tracking_details(self, foglamp_url, event=None):
-        _connection = http.client.HTTPConnection(foglamp_url)
-        uri = '/foglamp/track'
-        if event:
-            uri += '?event={}'.format(event)
-        _connection.request("GET", uri)
-        r = _connection.getresponse()
-        assert 200 == r.status
-        r = r.read().decode()
-        jdoc = json.loads(r)
-        return jdoc
 
     @pytest.fixture
     def start_south_north(self, reset_and_start_foglamp, add_south, enable_schedule, remove_directories,
@@ -157,14 +146,14 @@ class TestE2eFilterFFTThreshold:
         if not skip_verify_north_interface:
             self._verify_egress(read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries)
 
-        tracking_details = self.get_asset_tracking_details(foglamp_url, "Ingest")
+        tracking_details = utils.get_asset_tracking_details(foglamp_url, "Ingest")
         assert len(tracking_details["track"]), "Failed to track Ingest event"
         tracked_item = tracking_details["track"][0]
         assert "playfilter" == tracked_item["service"]
         assert "e2e_fft_threshold FFT" == tracked_item["asset"]
         assert "playback" == tracked_item["plugin"]
 
-        # TODO: Add asset tracker entry for fft and threshold filters
+        # TODO: FOGL:3440,FOGL:3441 Add asset tracker entry for fft and threshold filters
         # tracking_details = self.get_asset_tracking_details(foglamp_url, "Filter")
         # assert len(tracking_details["track"]), "Failed to track Ingest event"
         # tracked_item = tracking_details["track"][0]
@@ -173,7 +162,7 @@ class TestE2eFilterFFTThreshold:
         # assert "FFT Filter" == tracked_item["plugin"]
 
         if not skip_verify_north_interface:
-            egress_tracking_details = self.get_asset_tracking_details(foglamp_url,"Egress")
+            egress_tracking_details = utils.get_asset_tracking_details(foglamp_url,"Egress")
             assert len(egress_tracking_details["track"]), "Failed to track Egress event"
             tracked_item = egress_tracking_details["track"][0]
             assert "NorthReadingsTo_PI" == tracked_item["service"]
