+++ tests/system/python/e2e/test_e2e_kafka.py
@@ -51,17 +51,6 @@ class TestE2EKafka:
         jdoc = json.loads(r)
         return utils.serialize_stats_map(jdoc)
 
-    def get_asset_tracking_details(self, foglamp_url, event=None):
-        _connection = http.client.HTTPConnection(foglamp_url)
-        uri = '/foglamp/track'
-        if event:
-            uri += '?event={}'.format(event)
-        _connection.request("GET", uri)
-        r = _connection.getresponse()
-        assert 200 == r.status
-        r = r.read().decode()
-        jdoc = json.loads(r)
-        return jdoc
 
     def _prepare_template_reading_from_fogbench(self):
         """ Define the template file for fogbench readings """
@@ -176,7 +165,7 @@ class TestE2EKafka:
         if not skip_verify_north_interface:
             self._read_from_kafka(kafka_host, kafka_rest_port, kafka_topic)
 
-        tracking_details = self.get_asset_tracking_details(foglamp_url, "Ingest")
+        tracking_details = utils.get_asset_tracking_details(foglamp_url, "Ingest")
         assert len(tracking_details["track"]), "Failed to track Ingest event"
         tracked_item = tracking_details["track"][0]
         assert "coap" == tracked_item["service"]
@@ -184,7 +173,7 @@ class TestE2EKafka:
         assert "coap" == tracked_item["plugin"]
 
         if not skip_verify_north_interface:
-            egress_tracking_details = self.get_asset_tracking_details(foglamp_url,"Egress")
+            egress_tracking_details = utils.get_asset_tracking_details(foglamp_url,"Egress")
             assert len(egress_tracking_details["track"]), "Failed to track Egress event"
             tracked_item = egress_tracking_details["track"][0]
             assert "NorthReadingsTo{}".format(NORTH_PLUGIN_NAME) == tracked_item["service"]
