+++ tests/system/python/e2e/test_e2e_vary_asset_http_pi.py
@@ -42,17 +42,6 @@ class TestE2EAssetHttpPI:
         jdoc = json.loads(r)
         return utils.serialize_stats_map(jdoc)
 
-    def get_asset_tracking_details(self, foglamp_url, event=None):
-        _connection = http.client.HTTPConnection(foglamp_url)
-        uri = '/foglamp/track'
-        if event:
-            uri += '?event={}'.format(event)
-        _connection.request("GET", uri)
-        r = _connection.getresponse()
-        assert 200 == r.status
-        r = r.read().decode()
-        jdoc = json.loads(r)
-        return jdoc
 
     def _verify_egress(self, read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries, asset_name,
                        sensor_data, sensor_data_2):
@@ -190,7 +179,7 @@ class TestE2EAssetHttpPI:
             self._verify_egress(read_data_from_pi, pi_host, pi_admin, pi_passwd, pi_db, wait_time, retries, asset_name,
                                 sensor_data, sensor_data_2)
 
-        tracking_details = self.get_asset_tracking_details(foglamp_url, "Ingest")
+        tracking_details = utils.get_asset_tracking_details(foglamp_url, "Ingest")
         assert len(tracking_details["track"]), "Failed to track Ingest event"
         tracked_item = tracking_details["track"][0]
         assert "http_south" == tracked_item["service"]
@@ -198,7 +187,7 @@ class TestE2EAssetHttpPI:
         assert "http_south" == tracked_item["plugin"]
 
         if not skip_verify_north_interface:
-            egress_tracking_details = self.get_asset_tracking_details(foglamp_url,"Egress")
+            egress_tracking_details = utils.get_asset_tracking_details(foglamp_url,"Egress")
             assert len(egress_tracking_details["track"]), "Failed to track Egress event"
             tracked_item = egress_tracking_details["track"][0]
             assert "NorthReadingsToPI" == tracked_item["service"]
